+++ packages/koishi/src/utils/render.ts
@@ -2,6 +2,20 @@ import { Context, h } from 'koishi'
 import { Config, FamilyGames } from '@/interface'
 import { ImgRender, getStatHtml } from 'steam-family-bot-core/render'
 
+let enable = false
+// import {} from 'koishi-plugin-puppeteer'
+
+async function init() {
+  try {
+    await import('koishi-plugin-puppeteer')
+    enable = true
+  } catch (e) {
+    console.error('koishi-plugin-puppeteer not installed, render is disable')
+  }
+}
+
+init()
+
 export class KoishiImgRender implements ImgRender {
   private ctx: Context
   private config: Config
@@ -15,18 +29,22 @@ export class KoishiImgRender implements ImgRender {
     onError?: () => void
   ) {
     const res = genHtml()
+    // @ts-ignore
     const buf = await this.ctx.puppeteer.render(res, async (page, next) => {
       onStart?.()
       await new Promise<void>((resolve, reject) => {
         setTimeout(resolve, 5000)
       })
-      return page
-        .$('body')
-        .then(next)
-        .catch((e) => {
-          onError?.()
-          return ''
-        })
+      return (
+        page
+          .$('body')
+          // @ts-ignore
+          .then(next)
+          .catch((e) => {
+            onError?.()
+            return ''
+          })
+      )
     })
     return buf
   }
@@ -34,6 +52,7 @@ export class KoishiImgRender implements ImgRender {
     token: string,
     onStart?: () => void
   ): Promise<any> {
+    // @ts-ignore
     const page = await this.ctx.puppeteer.page()
     await page.setViewport({
       width: 1920,
@@ -54,7 +73,7 @@ export class KoishiImgRender implements ImgRender {
       }, 20000)
     })
     const elm = await page.waitForSelector('#data-graph', { timeout: 20000 })
-    const buffer = await elm.screenshot({})
+    const buffer = await elm!.screenshot({})
     await page.close()
     const image = h.image(buffer, 'image/png')
     return image
@@ -66,5 +85,9 @@ export class KoishiImgRender implements ImgRender {
     onError?: () => void
   ): Promise<string> {
     return this._render(() => getStatHtml(games), onStart, onError)
+  }
+
+  isRenderEnable(): boolean {
+    return enable
   }
 }