+++ packages/koishi/tests/index.spec.ts
@@ -1,17 +1,11 @@
 // @ts-nocheck
 import { Context } from 'koishi'
 import mock from '@koishijs/plugin-mock'
-import cron from 'koishi-plugin-cron'
-import puppeteer from 'koishi-plugin-puppeteer'
+import console from '@koishijs/plugin-console'
+import * as sandbox from '@koishijs/plugin-sandbox'
+import * as echo from '@koishijs/plugin-echo'
 
-const app = new Context()
-app.plugin(mock)
-app.plugin(puppeteer)
-app.plugin({
-  ...cron,
-})
 // app.plugin(steamBot)
-
 // const client = app.mock.client('123')
 //
 // app.start()
@@ -21,3 +15,4 @@ app.plugin({
 // it('example 1', async () => {
 //   await client.shouldReply('slm 你好','你好')
 // })
+// test