+++ packages/koishi/CHANGELOG.md
@@ -1,5 +1,90 @@
 # koishi-plugin-steam-family-bot
 
+## 0.0.7
+
+### Patch Changes
+
+- 4f157bb: fix: add koishi peerDependencies
+- 1bd27e3: update docs & fix unsub command delete account
+- i18n & bug fix
+- e665e74: optional puppeteer deps
+- bde33d0: refactor: improve code structure
+- dc16917: fix: make koishi-plugin-puppeteer optional
+- 3268030: fix: fix session import
+- 9fe4021: refactor: improve code structure
+- c580188: i18n support
+- Updated dependencies [1bd27e3]
+- Updated dependencies
+- Updated dependencies [e665e74]
+- Updated dependencies [bde33d0]
+- Updated dependencies [dc16917]
+- Updated dependencies [3268030]
+- Updated dependencies [9fe4021]
+- Updated dependencies [c580188]
+  - steam-family-bot-core@0.0.7
+
+## 0.0.7-dev.7
+
+### Patch Changes
+
+- update docs & fix unsub command delete account
+- Updated dependencies
+  - steam-family-bot-core@0.0.7-dev.7
+
+## 0.0.7-dev.6
+
+### Patch Changes
+
+- fix: fix session import
+- Updated dependencies
+  - steam-family-bot-core@0.0.7-dev.6
+
+## 0.0.7-dev.5
+
+### Patch Changes
+
+- refactor: improve code structure
+- Updated dependencies
+  - steam-family-bot-core@0.0.7-dev.5
+
+## 0.0.7-dev.4
+
+### Patch Changes
+
+- refactor: improve code structure
+- Updated dependencies
+  - steam-family-bot-core@0.0.7-dev.4
+
+## 0.0.7-dev.3
+
+### Patch Changes
+
+- fix: make koishi-plugin-puppeteer optional
+- Updated dependencies
+  - steam-family-bot-core@0.0.7-dev.3
+
+## 0.0.7-dev.2
+
+### Patch Changes
+
+- fix: add koishi peerDependencies
+
+## 0.0.7-dev.1
+
+### Patch Changes
+
+- optional puppeteer deps
+- Updated dependencies
+  - steam-family-bot-core@0.0.7-dev.1
+
+## 0.0.7-dev.0
+
+### Patch Changes
+
+- i18n support
+- Updated dependencies
+  - steam-family-bot-core@0.0.7-dev.0
+
 ## 0.0.6
 
 ### Patch Changes